refactor(labtest): rename lab-standards to lab-indicators
Unify naming: lab-standards.ts -> lab-indicators.ts
